requirements.txt
requests
bs4
flask
google.generativeai
